mock_services/profile/pyproject.toml
[project]
name = "profile-service"
version = "0.1.0"
description = "A simple mock profile service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "opentelemetry-distro>=0.55b1",
    "opentelemetry-exporter-otlp>=1.34.1",
    "opentelemetry-instrumentation-asgi>=0.55b1",
    "opentelemetry-instrumentation-flask>=0.55b1",
    "opentelemetry-instrumentation-logging>=0.55b1",
]
